% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_sampler.R
\name{lda}
\alias{lda}
\title{Latent Dirichlet Allocation}
\usage{
lda(size, model, alpha)
}
\arguments{
\item{size}{An integer or vector of integers.}

\item{model}{A matrix of probabilities. Model[i, j] should be the probability of word j given topic i. The value can either be a probability, [0, 1], or a frequency, [0, inf).}

\item{alpha}{A numeric or vector of numerics. The parameter(s) to the topic dirichlet distribution. The distribution will be symmetric if alpha is a single value.}
}
\description{
Samples documents based on the generative process in Latent Dirichlet Allocation.
}
